# server/openapiv2/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "group.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/balance/api/delete/tag": {
      "post": {
        "operationId": "UserTagService_DeleteUserTag",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoDeleteUserTagResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoDeleteUserTagRequest"
            }
          }
        ],
        "tags": [
          "UserTagService"
        ]
      }
    },
...
    "/balance/api/get/tag": {
      "post": {
        "operationId": "UserTagService_GetUserTag",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoGetUserTagResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoGetUserTagRequest"
            }
          }
        ],
        "tags": [
          "UserTagService"
        ]
      }
    },
...
    "/balance/api/list/tag": {
      "post": {
        "operationId": "UserTagService_ListUserTag",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoListUserTagResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoListUserTagRequest"
            }
          }
        ],
        "tags": [
          "UserTagService"
        ]
      }
    },
...
    "/balance/api/register/tag": {
      "post": {
        "operationId": "UserTagService_RegisterUserTag",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoRegisterUserTagResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoRegisterUserTagRequest"
            }
          }
        ],
        "tags": [
          "UserTagService"
        ]
      }
    },
...
    "/balance/api/update/tag": {
      "post": {
        "operationId": "UserTagService_UpdateUserTag",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoUpdateUserTagResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoUpdateUserTagRequest"
            }
          }
        ],
        "tags": [
          "UserTagService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protoDeleteUserTagRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "sessoinId": {
          "type": "string"
        }
      }
    },
    "protoDeleteUserTagResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        }
      }
    },
...
    "protoGetUserTagRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "sessoinId": {
          "type": "string"
        }
      }
    },
    "protoGetUserTagResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "tagName": {
          "type": "string"
        },
        "hasGroup": {
          "type": "boolean"
        },
        "groupId": {
          "type": "string",
          "format": "int64"
        },
        "grantLimit": {
          "type": "string"
        }
      }
    },
...
    "protoListUserTagRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "sessoinId": {
          "type": "string"
        }
      }
    },
    "protoListUserTagResponse": {
      "type": "object",
      "properties": {
        "userTagList": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protoUserTagResponse"
          }
        }
      }
    },
...
    "protoRegisterUserTagRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "tagName": {
          "type": "string"
        },
        "hasGroup": {
          "type": "boolean"
        },
        "groupId": {
          "type": "string",
          "format": "int64"
        },
        "grantLimit": {
          "type": "string"
        },
        "sessoinId": {
          "type": "string"
        }
      }
    },
    "protoRegisterUserTagResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "tagName": {
          "type": "string"
        },
        "hasGroup": {
          "type": "boolean"
        },
        "groupId": {
          "type": "string",
          "format": "int64"
        },
        "grantLimit": {
          "type": "string"
        }
      }
    },
...
    "protoUpdateUserTagRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "sessoinId": {
          "type": "string"
        }
      }
    },
    "protoUpdateUserTagResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "tagName": {
          "type": "string"
        },
        "hasGroup": {
          "type": "boolean"
        },
        "groupId": {
          "type": "string",
          "format": "int64"
        },
        "grantLimit": {
          "type": "string"
        }
      }
    },
    "protoUserTagResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "tagName": {
          "type": "string"
        },
        "hasGroup": {
          "type": "boolean"
        },
        "groupId": {
          "type": "string",
          "format": "int64"
        },
        "grantLimit": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}